we're done with the location search and predictions. moved na rin yung mga files sa mga folders kung saan sila dapat nakalagay
