+++ lib/authenticationAccounts/authGate.dart
@@ -8,8 +8,8 @@ authenticated - main page
 */
 
 import 'package:flutter/material.dart';
-import 'package:pasada_passenger_app/selectionScreen.dart';
-import 'package:pasada_passenger_app/loginAccount.dart';
+import 'package:pasada_passenger_app/home/selectionScreen.dart';
+import 'package:pasada_passenger_app/authenticationAccounts/loginAccount.dart';
 import 'package:supabase_flutter/supabase_flutter.dart';
 
 class AuthGate extends StatelessWidget {
