+++ lib/authenticationAccounts/createAccount.dart
@@ -1,8 +1,8 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_svg/flutter_svg.dart';
 import 'package:flutter_animate/flutter_animate.dart';
-import 'package:pasada_passenger_app/main.dart';
-import 'package:pasada_passenger_app/createAccountCred.dart';
+import 'package:pasada_passenger_app/home/main.dart';
+import 'package:pasada_passenger_app/authenticationAccounts/createAccountCred.dart';
 
 void main() => runApp(const CreateAccountPage());
