+++ lib/database/databaseSetup.dart
@@ -1,14 +1,14 @@
 import 'package:postgres/postgres.dart';
 import 'package:flutter/foundation.dart';
-import 'package:pasada_passenger_app/main.dart';
-import 'package:pasada_passenger_app/createAccountCred.dart';
-import 'package:pasada_passenger_app/activityScreen.dart';
-import 'package:pasada_passenger_app/createAccount.dart';
-import 'package:pasada_passenger_app/homeScreen.dart';
-import 'package:pasada_passenger_app/loginAccount.dart';
-import 'package:pasada_passenger_app/notificationScreen.dart';
-import 'package:pasada_passenger_app/profileSettingsScreen.dart';
-import 'package:pasada_passenger_app/settingsScreen.dart';
+import 'package:pasada_passenger_app/home/main.dart';
+import 'package:pasada_passenger_app/authenticationAccounts/createAccountCred.dart';
+import 'package:pasada_passenger_app/home/activityScreen.dart';
+import 'package:pasada_passenger_app/authenticationAccounts/createAccount.dart';
+import 'package:pasada_passenger_app/home/homeScreen.dart';
+import 'package:pasada_passenger_app/authenticationAccounts/loginAccount.dart';
+import 'package:pasada_passenger_app/home/notificationScreen.dart';
+import 'package:pasada_passenger_app/home/profileSettingsScreen.dart';
+import 'package:pasada_passenger_app/home/settingsScreen.dart';
 import 'package:postgres/postgres.dart';
 import 'dart:convert';
 import 'package:http/http.dart' as http;
