+++ lib/home/homeScreen.dart
@@ -3,8 +3,8 @@ import 'package:flutter/material.dart';
 // import 'package:url_launcher/url_launcher.dart';
 import 'package:location/location.dart';
 import 'package:google_maps_flutter/google_maps_flutter.dart';
-import 'package:pasada_passenger_app/mapScreen.dart';
-import 'locationSearchScreen.dart';
+import 'package:pasada_passenger_app/location/mapScreen.dart';
+import '../location/locationSearchScreen.dart';
 // import 'package:flutter_dotenv/flutter_dotenv.dart';
 // import 'package:pasada_passenger_app/notificationScreen.dart';
 // import 'package:pasada_passenger_app/activityScreen.dart';
