+++ lib/home/notificationScreen.dart
@@ -1,19 +1,19 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_svg/flutter_svg.dart';
-import 'package:pasada_passenger_app/homeScreen.dart';
+import 'package:pasada_passenger_app/home/homeScreen.dart';
 import 'package:url_launcher/url_launcher.dart';
 import 'package:location/location.dart';
 import 'package:google_maps_flutter/google_maps_flutter.dart';
-import 'package:pasada_passenger_app/notificationScreen.dart';
-import 'package:pasada_passenger_app/activityScreen.dart';
-import 'package:pasada_passenger_app/profileSettingsScreen.dart';
-import 'package:pasada_passenger_app/settingsScreen.dart';
-import 'package:pasada_passenger_app/homeScreen.dart';
+import 'package:pasada_passenger_app/home/notificationScreen.dart';
+import 'package:pasada_passenger_app/home/activityScreen.dart';
+import 'package:pasada_passenger_app/home/profileSettingsScreen.dart';
+import 'package:pasada_passenger_app/home/settingsScreen.dart';
+import 'package:pasada_passenger_app/home/homeScreen.dart';
 
-void main() => runApp(const SettingsScreen());
+void main() => runApp(const NotificationScreen());
 
-class SettingsScreen extends StatelessWidget {
-  const SettingsScreen({super.key});
+class NotificationScreen extends StatelessWidget {
+  const NotificationScreen({super.key});
 
   @override
   Widget build(BuildContext context) {
@@ -21,11 +21,11 @@ class SettingsScreen extends StatelessWidget {
       title: 'Pasada',
       debugShowCheckedModeBanner: false,
       theme: ThemeData(
-        scaffoldBackgroundColor: const Color(0xFFF2F2F2),
+        scaffoldBackgroundColor: const Color(0XFFF2F2F2),
         fontFamily: 'Inter',
         useMaterial3: true,
       ),
-      home: const SettingsScreenStateful(title:  'Pasada'),
+      home: const NotifScreenStateful(title: 'Pasada'),
       routes: <String, WidgetBuilder>{
 
       },
@@ -33,16 +33,17 @@ class SettingsScreen extends StatelessWidget {
   }
 }
 
-class SettingsScreenStateful extends StatefulWidget {
-  const SettingsScreenStateful({super.key, required this.title});
+class NotifScreenStateful extends StatefulWidget {
+  const NotifScreenStateful({super.key, required this.title});
 
   final String title;
 
   @override
-  State<SettingsScreenStateful> createState() => SettingsScreenPageState();
+  State<NotifScreenStateful> createState() => NotifScreenPageState();
 }
 
-class SettingsScreenPageState extends State<SettingsScreenStateful> {
+class NotifScreenPageState extends State<NotifScreenStateful> {
+
   @override
   Widget build(BuildContext context) {
     final screenHeight = MediaQuery.of(context).size.height;
@@ -56,7 +57,7 @@ class SettingsScreenPageState extends State<SettingsScreenStateful> {
               left: screenWidth * 0.05,
               right: screenWidth * 0.18,
               child: Text(
-                'Settings',
+                'Notifications',
                 style: TextStyle(
                   fontSize: 24,
                   fontWeight: FontWeight.w700,
