+++ lib/home/profileSettingsScreen.dart
@@ -1,19 +1,19 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_svg/flutter_svg.dart';
-import 'package:pasada_passenger_app/homeScreen.dart';
+import 'package:pasada_passenger_app/home/homeScreen.dart';
 import 'package:url_launcher/url_launcher.dart';
 import 'package:location/location.dart';
 import 'package:google_maps_flutter/google_maps_flutter.dart';
-import 'package:pasada_passenger_app/notificationScreen.dart';
-import 'package:pasada_passenger_app/activityScreen.dart';
-import 'package:pasada_passenger_app/profileSettingsScreen.dart';
-import 'package:pasada_passenger_app/settingsScreen.dart';
-import 'package:pasada_passenger_app/homeScreen.dart';
+import 'package:pasada_passenger_app/home/notificationScreen.dart';
+import 'package:pasada_passenger_app/home/activityScreen.dart';
+import 'package:pasada_passenger_app/home/profileSettingsScreen.dart';
+import 'package:pasada_passenger_app/home/settingsScreen.dart';
+import 'package:pasada_passenger_app/home/homeScreen.dart';
 
-void main() => runApp(const NotificationScreen());
+void main() => runApp(const ProfileScreen());
 
-class NotificationScreen extends StatelessWidget {
-  const NotificationScreen({super.key});
+class ProfileScreen extends StatelessWidget {
+  const ProfileScreen({super.key});
 
   @override
   Widget build(BuildContext context) {
@@ -21,11 +21,11 @@ class NotificationScreen extends StatelessWidget {
       title: 'Pasada',
       debugShowCheckedModeBanner: false,
       theme: ThemeData(
-        scaffoldBackgroundColor: const Color(0XFFF2F2F2),
+        scaffoldBackgroundColor: const Color(0xFFF2F2F2),
         fontFamily: 'Inter',
         useMaterial3: true,
       ),
-      home: const NotifScreenStateful(title: 'Pasada'),
+      home: const ProfileScreenStateful(title: 'Pasada'),
       routes: <String, WidgetBuilder>{
 
       },
@@ -33,16 +33,16 @@ class NotificationScreen extends StatelessWidget {
   }
 }
 
-class NotifScreenStateful extends StatefulWidget {
-  const NotifScreenStateful({super.key, required this.title});
+class ProfileScreenStateful extends StatefulWidget {
+  const ProfileScreenStateful({super.key, required this.title});
 
   final String title;
 
   @override
-  State<NotifScreenStateful> createState() => NotifScreenPageState();
+  State<ProfileScreenStateful> createState() => ProfileScreenPageState();
 }
 
-class NotifScreenPageState extends State<NotifScreenStateful> {
+class ProfileScreenPageState extends State<ProfileScreenStateful> {
 
   @override
   Widget build(BuildContext context) {
@@ -57,7 +57,7 @@ class NotifScreenPageState extends State<NotifScreenStateful> {
               left: screenWidth * 0.05,
               right: screenWidth * 0.18,
               child: Text(
-                'Notifications',
+                'Profile',
                 style: TextStyle(
                   fontSize: 24,
                   fontWeight: FontWeight.w700,
