+++ lib/home/selectionScreen.dart
@@ -3,11 +3,11 @@ import 'package:flutter_svg/flutter_svg.dart';
 import 'package:url_launcher/url_launcher.dart';
 import 'package:location/location.dart';
 import 'package:google_maps_flutter/google_maps_flutter.dart';
-import 'package:pasada_passenger_app/notificationScreen.dart';
-import 'package:pasada_passenger_app/activityScreen.dart';
-import 'package:pasada_passenger_app/profileSettingsScreen.dart';
-import 'package:pasada_passenger_app/settingsScreen.dart';
-import 'package:pasada_passenger_app/homeScreen.dart';
+import 'package:pasada_passenger_app/home/notificationScreen.dart';
+import 'package:pasada_passenger_app/home/activityScreen.dart';
+import 'package:pasada_passenger_app/home/profileSettingsScreen.dart';
+import 'package:pasada_passenger_app/home/settingsScreen.dart';
+import 'package:pasada_passenger_app/home/homeScreen.dart';
 
 class selectionScreen extends StatefulWidget {
   const selectionScreen({super.key});
